api/requirements.txt
anyio==3.6.2
azure-cognitiveservices-speech==1.25.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
deepl==1.13.0
fastapi==0.89.1
ffprobe-python==1.0.3
google-api-core==2.11.0
google-api-python-client==2.76.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
google-cloud-texttospeech==2.14.1
googleapis-common-protos==1.58.0
grpcio==1.51.1
grpcio-status==1.51.1
h11==0.14.0
httplib2==0.21.0
httptools==0.5.0
idna==3.4
langcodes==3.3.0
language-data==1.1
marisa-trie==0.7.8
numpy==1.24.2
oauthlib==3.2.2
pandas==1.5.3
proto-plus==1.22.2
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.4
pydub==0.25.1
pyparsing==3.0.9
pyrubberband==0.3.0
PySoundFile==0.9.0.post1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.14
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
